deps: Make `const-random` optional via feature
Previously, `const-random` was a required dependency, even though it was
only used in no_std environments.  The default for the crate is std.
This change introduces a new feature "const-random" that brings in the
`const-random` dependency for no_std environments.  Because there isn't
a way in cargo to make two features mutually exclusive, a `compile_err`
is also added to ensure the crate can't compile if the `std` and
`const-random` features are both enabled.